apps/api/requirements/dev.txt
-r base.txt
pytest
pytest-flask
black
ruff
mypy
